requirements.txt
selenium
webdriver_manager
pytest
faker
allure-pytest